Added center alignment option
`TRAY_ALIGN_CENTER` has been added to the "registry setting overrides"
section. Using this setting will skip the taskbar position check and use
a center-top anchor for the menu instead, which works for all taskbar
positions. Note that the menu will still be stopped by the edge of the
screen, or if you have the icon in your overflow menu, the edge of your
taskbar. The menu will have a strange alignment in these cases.