Revamped path settings, added AppData options
The log and history files can now be saved to the local AppData folder
under the "instant-replay-suite" directory. Settings for specifying the
names or absolute paths of the log file, icon file, history file, and
resource directory have been added. These settings will resolve the
paths regardless of whether the setting is a relative or absolute path.