Added undo ability for trimming and concatenating
Default hotkey is Alt + U. Trims and concatenations now write to a text
file (for persistence between sessions) called `UNDO_LIST_PATH`,
detailing the name of the backup file(s) that were altered, the path the
altered version(s) were saved to, and a short description of the edit.
When undoing, this text file is read back and used to remove the edited
clip and restore the backed up version before refreshing the Clip
objects (unless they were restored as cached strings). The menu-action
for this is called "undo". Both the basic and advanced menus include
this action by default.

Trims and concatenations both have unique sound alerts for undoing them.
Also, `get_clip` has been updated to support a `path` parameter.