Simplified clip-checking system
`check_for_clips_thread` now uses nested `os.llistdir` calls instead of
an `os.walk` call, and ONLY the base files of the base subfolders in
`VIDEO_FOLDER` are looked at. This is to ensure that only files
that could have been created by Instant Replay are looked at, reducing
occurances of user-made videos being detected (for now).

This change also fixes `IGNORE_VIDEOS_IN_THESE_FOLDERS` not working
when the video in question was in a subfolder of an ignored folder.