Made `CLIP_BUFFER` more intuitive, lowered minimum
`CLIP_BUFFER` is supposed to cover every item that appears in the
menu, plus one extra for concatenating the final visible item. This
extra clip was not accounted for in `CLIP_BUFFER` itself, but instead
every calculation related to the buffer and caching/uncaching, which
was very unintuitive. This is no longer the case. All calculations
have been updated accordingly.

Additionally, the minimum `CLIP_BUFFER` has
been lowered from 5 to 2 (technically 6 to 3).