Added external config file/library, new defaults
ConfigParseBetter (my own personal library) has been added (and has
received massive updates itself for this script). It allows comments,
case sensitive settings, and simplified but robust syntax, as well as
a whole host of other features that are not used here at all. The config
file will only be created if one does not already exist, though this
will be changed in the future to always rewrite the config file.

Changed defaults:
 - MAX_BACKUPS: 5 -> 10
 - TRAY_RECENT_CLIP_COUNT: 5 -> 10
 - RENAME_COUNT_PADDED_ZEROS: 0 -> 1

 The config file is called "config.settings.ini". Settings related to
 custom log file paths have been removed for now due to complications
 with future commits.