Added hint for `SAVE_BACKUPS_TO_VIDEO_FOLDER`
To avoid an unnecessary message box from appearing on first launch for
some users, `VIDEO_FOLDER` is now detected *before* reading the config
(and the other registry settings), so we can check if it's on the same
drive as us. If it isn't, we set a hint called `BACKUP_FOLDER_HINT` to
True, which is used as the fallback for `SAVE_BACKUPS_TO_VIDEO_FOLDER`.

This way, if the user hasn't changed that setting manually and doesn't
have custom `VIDEO_FOLDER` or `BACKUP_FOLDER` paths set, we can default
`BACKUP_FOLDER` to be inside `VIDEO_FOLDER`, preventing the issue from
ever coming up in the first place.