Compiling-related touchups/additions
All instances of `exit` now use the compile-safe `sys.exit`, the
constant `IS_COMPILED` has been added, `MEDIAINFO_DLL_PATH` has been
added for specifying a path to "MediaInfo.dll" in a separate folder,
and `LOG_PATH` is "InstantReplaySuite.log" while compiled, for now.

Another constant, `BIN_FOLDER`, has been added, and currently defaults
to "PIL." Normally we would hide all our excess files/folders in a
generic "bin" folder after compiling, but hiding the PIL folder this
way is quite difficult due to it requiring a from-import. So instead,
we'll be using the PIL folder itself as our "bin" folder and dumping
everything in there.

Additionally, we'll be hiding a secret copy of our default icon with
a fake name in our bin folder, just to use as a backup in case the user
deletes the real icon. The backup will be named "libico.dll." If we're
running directly from the script however, a warning will be shown and
the script will exit with exit code 3 (missing file).

Also, an alias for `pymediainfo.MediaInfo.parse` has been added.