Removed `PIL`, added ability to use .exe icon
`Icon._assert_icon_handle` is now overridden to support strings for
`self._icon` rather than `PIL.Image.Image`, as its only usage in
`pystray` was to make sure a given icon was valid. Due to the heaviness
of importing/compiling `PIL`, it's worth removing, even though it means
you can no longer use regular image files as icons without converting
them first. This may be rolled back in the future.

If the given icon isn't valid, we'll use `win32api.LoadResource` and
`win32gui.CreateIconFromResource` to take the icon embedded in our
executable, if we're compiled. These libraries are already included in
our compilation, so it costs essentially nothing to use them here.

For our current icon, index 4 returns the best icon from the executable,
even at different scales/resolutions. This may change if the embedded
icon is changed. If for whatever reason our embedded icon still cannot
be used (or we're not compiled), a warning in shown and we exit.

`tray_icon._assert_icon_handle` is called manually before running, so
that exiting works if needed. The `ICON` setting has been renamed to
`CUSTOM_ICON` for clarity, and now defaults to nothing, as we want
to use the .exe icon by default so we don't have to worry about
replacing it during an update.

Additionally, minor refactoring has been done to our `Icon` class.