Edited resources are now ignored during updates
Compiled folders will include a "!defaults.txt" file in their resource
folder detailing the default resources and their sizes. This is used
to detect edited files before an update so we can tell our updater
not to overwrite the edited files. Deleted resources are also detected,
and are treated the same as edited resources for now.

`RESOURCE_FOLDER` is no longer editable (for now) to simplify things.