Improved auto-updating robustness
The lock file is now created in our main script and passed to
`update.check_for_update` to ensure the updater doesn't open
faster than we can exit our script.

Additionally, missing config files no longer get written if we're about
to restart for an update or if the missing-FFmpeg prompt is rejected.