Don't overwrite config if modified while running
Returns the MD5 hash of the config file (if present) from
`verify_config_files` and compares it against the MD5 hash of the same
file on exit to see if it's been modified. If it has, then the config is
not overwritten, avoiding the situation of editing a setting, closing
the program, and having the setting revert back.

Adds the constant `CONFIG_HASH` and imports the `hashlib` library.
The `CONFIG_PATH` constant is no longer deleted.