Added microphone/system audio track merging
`AutoCutter.merge_clip_audio_tracks` has been added, which merges all
existing audio tracks in a given clip into one. This is meant for
clips saved with the system and microphone audio on separate tracks.
This function works and is implemented similarly to the existing edit
methods. The default hotkey is `Alt + M` and it's been added to the
bottom of each recent clip's submenu.

`get_audio_tracks` has been added to get the audio tracks for a clip
at a given `path`, similar to `get_video_duration`, which has been
refactored.