Added history encoding & file-scanning timeout
The history file is now encoded and decoded (with some extra measures
taken) to fix errors with wacky silly characters in filenames.

`self.check_for_clips_thread()` now employs a retroactive timeout system
for scans - giving up on .mp4 files with a duration of 0 if it's been 10
seconds since the file was created. This not only adds a timeout system
to automatic scans, but allows manual scans to see if a file is already
a lost cause or if we should still wait for it.