Fixed trimming performance/audio desync
Trimming now seeks before specifying the input, which has drastically
different results for some idiotic reason. This fixes audio desync on
some systems, and ensures consistently-instant trims in all
circumstances.

Also, `ffmpeg()` has been adjusted to use `subprocess.STARTUPINFO` to
hide the window while compiled (instead of relying on `shell=True`) and
to expect "%in" in the `cmd` if `infile` is provided (to be slightly
more inline with how I do it in PyPlayer).