Simplified and improved how `?count` works
The count that appears in clip titles no longer tries to smartly
calculate itself to be as low (but consistent) as possible, and instead
remembers the number of clips saved for the given name template (i.e.
if your titles are specific down to the hour, it will track the number
of clips saved for that game in that hour, since they will share the
same name template). These counts are then saved to a new file
("count.txt" by default), which allows IRS to be restarted and
still accurately maintain the latest counts.

Old templates are cleared when new ones are created for the same game,
allowing memory usage to be low while tracking the clip counts, even
when the name templates are accurate to the second.

Adds `OLD_COUNTS_PATH`, `TEMPLATE_COUNTS`, and `TEMPLATE_FOLDERS`.