Added new TTS alerts/errors, improved length check
Added individual length alerts for every number from 1-20. Also added
a "clip is too short" alert when trying to trim a clip that is already
shorter than your desired length (this check is also now includes a
a 0.01 margin of error in case of FFmpeg oddities).

Additionally, entering a trim of length 0 will play a "trim cancelled"
alert.